Enforce required parameters for Telegram methods
verify_arguments prevents nil values from being sent thus enforces
required parameters for the methods.

grab_arguments is a function that takes all values of the parameters
(including the splat parameter) and make all of them into a single Hash
that could be used to pass parameters to the request method.

It's made to prevent constant repetition of the same lines of code.